refactor(phone): the diff and the threads become panes, not screens
Both are about to be read in two places: as a segment of the review, and
as the route a notification points at. Two implementations of that would
drift, and the drift would be a deep link opening a screen that behaves
unlike the one beside it.

So the screens move to `src/review/` as components taking their number
and root as props, and the routes become the four lines that read the
parameters and render them. Nothing else changes: same state, same
fetches, same behaviour.

The two source-scanning locks moved with the code they are about — the
tap floor's 22-point diff row, and the checks that the diff is a
windowed list whose row is called rather than mounted as a fresh
component type. They failed on the move, which is what they are for.
